Fix int32 truncation in tbe_input_combine offsets
Accumulate combined offsets in int64 to avoid overflow past
INT32_MAX; read per-table last offset via toLong(). Sync the
abstract/meta registration to int64 and align the reference
oracle accumulators (offsets_starts/offsets_accs) to int64.
Extract a shared combined_offsets allocation helper, guard
padding_fused offsets_size against out-of-bounds writes, and
extend tests with large-offset regression coverage for both ops.

+++ fbgemm_gpu/fbgemm_gpu/sparse_ops.py
@@ -383,7 +383,7 @@ def tbe_input_combine_abstract(
             need_weight = True
     total_offsets = torch.library.get_ctx().new_dynamic_size()
     combined_indices = indices_list[0].new_empty([total_indices], dtype=torch.int)
-    combined_offsets = offsets_list[0].new_empty([total_offsets], dtype=torch.int)
+    combined_offsets = offsets_list[0].new_empty([total_offsets], dtype=torch.long)
     if need_weight:
         combined_weights = per_sample_weights[0].new_empty(
             [total_indices], dtype=torch.float
